refactor: move Pointing type to web/lib/types.ts
Per CLAUDE.md convention, DB-shaped types live in web/lib/types.ts. Pointing
mirrors the JSONB structure stored in observations.pointings, so it belongs
there rather than in the server-action file.
