+++ web/lib/actions/programs.ts
@@ -1,6 +1,7 @@
 'use server';
 
 import { createClient } from '@/lib/supabase/server';
+import type { Pointing } from '@/lib/types';
 
 export interface ProgramOverview {
   slug: string;
@@ -14,24 +15,6 @@ export interface ProgramOverview {
   fields: string[];
   observations: string[];
   jwst_pids: number[];
-}
-
-export interface Pointing {
-  msametid: number;
-  msametfl: string;
-  ra_center: number;
-  dec_center: number;
-  pa_aper: number;
-  gratings: string[];
-  filters: string[];
-  jwst_program: number;
-  jwst_obs_ids: string[];
-  n_exposures: number;
-  n_dithers: number;
-  exptime_total: number;
-  date_obs_start: string;
-  date_obs_end: string;
-  footprint: number[][][]; // 4 quadrants × 4 corners × [ra, dec]
 }
 
 export interface ObservationStat {
